checkers/nodbrary/requirements.txt
Faker
beautifulsoup4
pycryptodome